initial refactoring, add breadcrumbs and grass visualization
