+++ gym_minigrid/__init__.py
@@ -1,5 +1,8 @@
 # Import the envs module so that envs register themselves
 import gym_minigrid.envs
 
+# Import shared constants
+import gym_minigrid.constants
+
 # Import wrappers so it's accessible when installing with pip
 import gym_minigrid.wrappers
