+++ gym_minigrid/envs/__init__.py
@@ -19,3 +19,4 @@ from gym_minigrid.envs.crossing import *
 from gym_minigrid.envs.lavagap import *
 from gym_minigrid.envs.dynamicobstacles import *
 from gym_minigrid.envs.distshift import *
+from gym_minigrid.envs.grass import *
