+++ gym_minigrid/envs/gotodoor.py
@@ -1,4 +1,4 @@
-from gym_minigrid.minigrid import *
+from gym_minigrid.minigrid_env import *
 from gym_minigrid.register import register
 
 class GoToDoorEnv(MiniGridEnv):
