+++ gym_minigrid/envs/lockedroom.py
@@ -1,5 +1,5 @@
 from gym import spaces
-from gym_minigrid.minigrid import *
+from gym_minigrid.minigrid_env import *
 from gym_minigrid.register import register
 
 class Room:
