+++ gym_minigrid/envs/redbluedoors.py
@@ -1,4 +1,4 @@
-from gym_minigrid.minigrid import *
+from gym_minigrid.minigrid_env import *
 from gym_minigrid.register import register
 
 class RedBlueDoorEnv(MiniGridEnv):
