+++ gym_minigrid/minigrid_env.py
@@ -1,0 +1,737 @@
+import hashlib
+from enum import IntEnum
+
+from gym.utils import seeding
+from gym.vector.utils import spaces
+
+from .world_obj import *
+from abc import ABC, abstractmethod
+from gym_minigrid.grid import Grid
+
+
+class MiniGridEnv(ABC):  # (gym.Env):
+
+    """
+    2D grid world game environment
+    """
+
+    # TODO removing 'human' mode as it doesn't jive with docker
+
+    metadata = {
+        'render.modes': ['rgb_array'],
+        'video.frames_per_second': 10
+    }
+
+    # Enumeration of possible actions
+    class Actions(IntEnum):
+        # Turn left, turn right, move forward
+        left = 0
+        right = 1
+        forward = 2
+
+        # Pick up an object
+        pickup = 3
+        # Drop an object
+        drop = 4
+        # Toggle/activate an object
+        toggle = 5
+
+        # Done completing task
+        done = 6
+
+    def __init__(
+            self,
+            grid_size=None,
+            width=None,
+            height=None,
+            max_steps=100,
+            see_through_walls=False,
+            seed=1337,
+            agent_view_size=7
+    ):
+        self.mission = None
+        self.grid = None
+        self.step_count = 0
+
+        self.carrying = None
+        self.np_random = None
+
+        # Can't set both grid_size and width/height
+        if grid_size:
+            assert not width and not height
+            width = grid_size
+            height = grid_size
+
+        # Action enumeration for this environment
+        self.actions = MiniGridEnv.Actions
+
+        # Actions are discrete integer values
+        self.action_space = spaces.Discrete(len(self.actions))
+
+        # Number of cells (width and height) in the agent view
+        assert agent_view_size % 2 == 1
+        assert agent_view_size >= 3
+        self.agent_view_size = agent_view_size
+
+        # Observations are dictionaries containing an
+        # encoding of the grid and a textual 'mission' string
+        self.observation_space = spaces.Box(
+            low=0,
+            high=255,
+            shape=(self.agent_view_size, self.agent_view_size, 3),
+            dtype='uint8'
+        )
+        self.observation_space = spaces.Dict({
+            'image': self.observation_space
+        })
+
+        # Range of possible rewards
+        self.reward_range = (0, 1)
+
+        #
+        # Range of penalties
+        self.penalty_range = (0, 1)
+
+        # Window to use for human rendering mode  # TODO disable this as it's useless in docker
+        # self.window = None
+
+        # Environment configuration
+        self.width = width
+        self.height = height
+        self.max_steps = max_steps
+        self.see_through_walls = see_through_walls
+
+        # Current position and direction of the agent
+        self.agent_pos = None
+        self.agent_dir = None
+
+        # Initialize the RNG
+        self.seed(seed=seed)
+
+        # breadcrumbs to show agent's path through the grid
+        self.breadcrumbs = np.zeros(shape=(self.width, self.height), dtype=np.bool)
+
+        # self.np_random = None
+
+        # Initialize the state
+        self.reset()
+
+    # -----------------------------------------------------
+    # gym.Env basic agent methods
+    # (step, reset, render, close, and seed)
+
+    def step(self, action):
+        self.step_count += 1
+
+        reward = 0
+        penalty = 0  # TODO
+        done = False
+
+        # Get the position in front of the agent
+        fwd_pos = self.front_pos
+
+        # Get the contents of the cell in front of the agent
+        fwd_cell = self.grid.get(*fwd_pos)
+
+        # Rotate left
+        if action == self.actions.left:
+            self.agent_dir -= 1
+            if self.agent_dir < 0:
+                self.agent_dir += 4
+
+        # Rotate right
+        elif action == self.actions.right:
+            self.agent_dir = (self.agent_dir + 1) % 4
+
+        # Move forward
+        elif action == self.actions.forward:
+            if not fwd_cell or fwd_cell.can_overlap():
+                # add last agent position to breadcrumbs
+                self.breadcrumbs[self.agent_pos[0], self.agent_pos[1]] = True
+                self.agent_pos = fwd_pos
+            if fwd_cell and fwd_cell.type == 'goal':
+                done = True
+                reward = self._reward()
+            if fwd_cell and fwd_cell.type == 'lava':
+                done = True
+            if fwd_cell and fwd_cell.type == 'grass':
+                # TODO this will incur a penalty
+                print("    stepping on grass now")
+
+        # Pick up an object
+        elif action == self.actions.pickup:
+            if fwd_cell and fwd_cell.can_pickup():
+                if self.carrying is None:
+                    self.carrying = fwd_cell
+                    self.carrying.cur_pos = np.array([-1, -1])
+                    self.grid.set(*fwd_pos, None)
+
+        # Drop an object
+        elif action == self.actions.drop:
+            if not fwd_cell and self.carrying:
+                self.grid.set(*fwd_pos, self.carrying)
+                self.carrying.cur_pos = fwd_pos
+                self.carrying = None
+
+        # Toggle/activate an object
+        elif action == self.actions.toggle:
+            if fwd_cell:
+                fwd_cell.toggle(self, fwd_pos)
+
+        # Done action (not used by default)
+        elif action == self.actions.done:
+            pass
+        else:
+            assert False, "unknown action"
+
+        if self.step_count >= self.max_steps:
+            done = True
+
+        obs = self.gen_obs()
+
+        return obs, reward, done, {}
+
+    def reset(self):
+
+        # Current position and direction of the agent
+        self.agent_pos = None
+        self.agent_dir = None
+
+        # clear breadcrumbs
+        self.breadcrumbs = np.zeros(shape=(self.width, self.height), dtype=np.bool)
+
+        # Generate a new random grid at the start of each episode
+        # To keep the same grid for each episode, call env.seed() with
+        # the same seed before calling env.reset()
+        self._gen_grid(self.width, self.height)
+
+        # These fields should be defined by _gen_grid
+        assert self.agent_pos is not None
+        assert self.agent_dir is not None
+
+        # Check that the agent doesn't overlap with an object
+        start_cell = self.grid.get(*self.agent_pos)
+        assert start_cell is None or start_cell.can_overlap()
+
+        # Item picked up, being carried, initially nothing
+        self.carrying = None
+
+        # Step count since episode start
+        self.step_count = 0
+
+        # Return first observation
+        obs = self.gen_obs()
+        return obs
+
+    def render(self, mode='rgb_array', close=False, highlight=True, tile_size=TILE_PIXELS):
+
+        """
+        Render the whole-grid rgb_array view
+        """
+
+        if close:
+            # if self.window:
+            #     self.window.close()
+            return
+
+        # if mode == 'human' and not self.window:
+        #     import gym_minigrid.window
+        #     self.window = gym_minigrid.window.Window('gym_minigrid')
+        #     self.window.show(block=False)
+
+        # Compute which cells are visible to the agent
+        _, vis_mask = self.gen_obs_grid()
+
+        # Compute the world coordinates of the bottom-left corner
+        # of the agent's view area
+        f_vec = self.dir_vec
+        r_vec = self.right_vec
+        top_left = self.agent_pos + f_vec * (self.agent_view_size - 1) - r_vec * (self.agent_view_size // 2)
+
+        # Mask of which cells to highlight
+        highlight_mask = np.zeros(shape=(self.width, self.height), dtype=np.bool)
+
+        # For each cell in the visibility mask
+        for vis_j in range(0, self.agent_view_size):
+            for vis_i in range(0, self.agent_view_size):
+                # If this cell is not visible, don't highlight it
+                if not vis_mask[vis_i, vis_j]:
+                    continue
+
+                # Compute the world coordinates of this cell
+                abs_i, abs_j = top_left - (f_vec * vis_j) + (r_vec * vis_i)
+
+                if abs_i < 0 or abs_i >= self.width:
+                    continue
+                if abs_j < 0 or abs_j >= self.height:
+                    continue
+
+                # Mark this cell to be highlighted
+                highlight_mask[abs_i, abs_j] = True
+
+        # Render the whole grid
+        img = self.grid.render(
+            tile_size,
+            self.agent_pos,
+            self.agent_dir,
+            highlight_mask=highlight_mask if highlight else None,
+            breadcrumbs=self.breadcrumbs
+        )
+
+        # if mode == 'human':
+        #     self.window.show_img(img)
+        #     self.window.set_caption(self.mission)
+
+        return img
+
+    def close(self):
+        #     # if self.window:
+        #     #     self.window.close()
+        return
+
+    def seed(self, seed=1337):
+        # Seed the random number generator
+        self.np_random, _ = seeding.np_random(seed)
+        return [seed]
+
+    # ----------------------------------------
+    # other methods
+
+    def hash(self, size=16):
+        """Compute a hash that uniquely identifies the current state of the environment.
+        :param size: Size of the hashing
+        """
+        sample_hash = hashlib.sha256()
+
+        to_encode = [self.grid.encode(), self.agent_pos, self.agent_dir]
+        for item in to_encode:
+            sample_hash.update(str(item).encode('utf8'))
+
+        return sample_hash.hexdigest()[:size]
+
+    @property
+    def steps_remaining(self):
+        return self.max_steps - self.step_count
+
+    def __str__(self):
+        """
+        Produce a pretty string of the environment's grid along with the agent.
+        A grid cell is represented by 2-character string, the first one for
+        the object and the second one for the color.
+        """
+
+        # Map of object types to short string
+        OBJECT_TO_STR = {
+            'wall': 'W',
+            'floor': 'F',
+            'door': 'D',
+            'key': 'K',
+            'ball': 'A',
+            'box': 'B',
+            'goal': 'G',
+            'lava': 'V',
+            'grass': 'R',
+        }
+
+        env_str = ''
+
+        for j in range(self.grid.height):
+
+            for i in range(self.grid.width):
+                if i == self.agent_pos[0] and j == self.agent_pos[1]:
+                    env_str += 2 * AGENT_DIR_TO_STR[self.agent_dir]
+                    continue
+
+                c = self.grid.get(i, j)
+
+                if not c:
+                    env_str += '  '
+                    continue
+
+                if c.type == 'door':
+                    if c.is_open:
+                        env_str += '__'
+                    elif c.is_locked:
+                        env_str += 'L' + c.color[0].upper()
+                    else:
+                        env_str += 'D' + c.color[0].upper()
+                    continue
+
+                env_str += OBJECT_TO_STR[c.type] + c.color[0].upper()
+
+            if j < self.grid.height - 1:
+                env_str += '\n'
+
+        return env_str
+
+    @abstractmethod
+    def _gen_grid(self, width, height):
+        pass
+        # assert False, "_gen_grid needs to be implemented by each environment"
+
+    def _reward(self):
+        """
+        Compute the reward to be given upon success
+        """
+        return 1 - 0.9 * (self.step_count / self.max_steps)
+
+    def _penalty(self):
+        # TODO compute the penalty (vs. setting state to done) for doing a bad thing
+        pass
+
+    def _rand_int(self, low, high):
+        """
+        Generate random integer in [low,high[
+        """
+        return self.np_random.randint(low, high)
+
+    def _rand_float(self, low, high):
+        """
+        Generate random float in [low,high[
+        """
+        return self.np_random.uniform(low, high)
+
+    def _rand_bool(self):
+        """
+        Generate random boolean value
+        """
+        return self.np_random.randint(0, 2) == 0
+
+    def _rand_elem(self, iterable):
+        """
+        Pick a random element in a list
+        """
+        lst = list(iterable)
+        idx = self._rand_int(0, len(lst))
+        return lst[idx]
+
+    def _rand_subset(self, iterable, num_elems):
+        """
+        Sample a random subset of distinct elements of a list
+        """
+        lst = list(iterable)
+        assert num_elems <= len(lst)
+        out = []
+        while len(out) < num_elems:
+            elem = self._rand_elem(lst)
+            lst.remove(elem)
+            out.append(elem)
+        return out
+
+    def _rand_color(self):
+        """
+        Generate a random color name (string)
+        """
+        return self._rand_elem(COLOR_NAMES)
+
+    def _rand_pos(self, xLow, xHigh, yLow, yHigh):
+        """
+        Generate a random (x,y) position tuple
+        """
+        return (
+            self.np_random.randint(xLow, xHigh),
+            self.np_random.randint(yLow, yHigh)
+        )
+
+    def place_obj(self,
+                  obj,
+                  top=None,
+                  size=None,
+                  reject_fn=None,
+                  max_tries=math.inf
+                  ):
+        """
+        Place an object at an empty position in the grid
+
+        :param obj:
+        :param max_tries:
+        :param top: top-left position of the rectangle where to place
+        :param size: size of the rectangle where to place
+        :param reject_fn: function to filter out potential positions
+        """
+
+        if top is None:
+            top = (0, 0)
+        else:
+            top = (max(top[0], 0), max(top[1], 0))
+
+        if size is None:
+            size = (self.grid.width, self.grid.height)
+
+        num_tries = 0
+
+        while True:
+            # This is to handle with rare cases where rejection sampling
+            # gets stuck in an infinite loop
+            if num_tries > max_tries:
+                raise RecursionError('rejection sampling failed in place_obj')
+
+            num_tries += 1
+
+            pos = np.array((
+                self._rand_int(top[0], min(top[0] + size[0], self.grid.width)),
+                self._rand_int(top[1], min(top[1] + size[1], self.grid.height))
+            ))
+
+            # Don't place the object on top of another object
+            if self.grid.get(*pos):
+                continue
+
+            # Don't place the object where the agent is
+            if np.array_equal(pos, self.agent_pos):
+                continue
+
+            # Check if there is a filtering criterion
+            if reject_fn and reject_fn(self, pos):
+                continue
+
+            break
+
+        self.grid.set(*pos, obj)
+
+        if obj is not None:
+            obj.init_pos = pos
+            obj.cur_pos = pos
+
+        return pos
+
+    def put_obj(self, obj, i, j):
+        """
+        Put an object at a specific position in the grid
+        """
+        self.grid.set(i, j, obj)
+        obj.init_pos = (i, j)
+        obj.cur_pos = (i, j)
+
+    def place_agent(
+            self,
+            top=None,
+            size=None,
+            rand_dir=True,
+            max_tries=math.inf
+    ):
+        """
+        Set the agent's starting point at an empty position in the grid
+        """
+        self.agent_pos = None
+        pos = self.place_obj(None, top, size, max_tries=max_tries)
+        self.agent_pos = pos
+        if rand_dir:
+            self.agent_dir = self._rand_int(0, 4)
+        return pos
+
+    @property
+    def dir_vec(self):
+        """
+        Get the direction vector for the agent, pointing in the direction
+        of forward movement.
+        """
+        assert 0 <= self.agent_dir < 4
+        return DIR_TO_VEC[self.agent_dir]
+
+    @property
+    def right_vec(self):
+        """
+        Get the vector pointing to the right of the agent.
+        """
+        dx, dy = self.dir_vec
+        return np.array((-dy, dx))
+
+    @property
+    def front_pos(self):
+        """
+        Get the position of the cell that is right in front of the agent
+        """
+        return self.agent_pos + self.dir_vec
+
+    def get_view_coords(self, i, j):
+        """
+        Translate and rotate absolute grid coordinates (i, j) into the
+        agent's partially observable view (sub-grid). Note that the resulting
+        coordinates may be negative or outside of the agent's view size.
+        """
+
+        ax, ay = self.agent_pos
+        dx, dy = self.dir_vec
+        rx, ry = self.right_vec
+
+        # Compute the absolute coordinates of the top-left view corner
+        sz = self.agent_view_size
+        hs = self.agent_view_size // 2
+        tx = ax + (dx * (sz - 1)) - (rx * hs)
+        ty = ay + (dy * (sz - 1)) - (ry * hs)
+
+        lx = i - tx
+        ly = j - ty
+
+        # Project the coordinates of the object relative to the top-left
+        # corner onto the agent's own coordinate system
+        vx = (rx * lx + ry * ly)
+        vy = -(dx * lx + dy * ly)
+
+        return vx, vy
+
+    def get_view_exts(self):
+        """
+        Get the extents of the square set of tiles visible to the agent
+        Note: the bottom extent indices are not included in the set
+        """
+
+        # Facing right
+        if self.agent_dir == 0:
+            topX = self.agent_pos[0]
+            topY = self.agent_pos[1] - self.agent_view_size // 2
+        # Facing down
+        elif self.agent_dir == 1:
+            topX = self.agent_pos[0] - self.agent_view_size // 2
+            topY = self.agent_pos[1]
+        # Facing left
+        elif self.agent_dir == 2:
+            topX = self.agent_pos[0] - self.agent_view_size + 1
+            topY = self.agent_pos[1] - self.agent_view_size // 2
+        # Facing up
+        elif self.agent_dir == 3:
+            topX = self.agent_pos[0] - self.agent_view_size // 2
+            topY = self.agent_pos[1] - self.agent_view_size + 1
+        else:
+            assert False, "invalid agent direction"
+
+        botX = topX + self.agent_view_size
+        botY = topY + self.agent_view_size
+
+        return topX, topY, botX, botY
+
+    def relative_coords(self, x, y):
+        """
+        Check if a grid position belongs to the agent's field of view, and returns the corresponding coordinates
+        """
+
+        vx, vy = self.get_view_coords(x, y)
+
+        if vx < 0 or vy < 0 or vx >= self.agent_view_size or vy >= self.agent_view_size:
+            return None
+
+        return vx, vy
+
+    def in_view(self, x, y):
+        """
+        check if a grid position is visible to the agent
+        """
+
+        return self.relative_coords(x, y) is not None
+
+    def agent_sees(self, x, y):
+        """
+        Check if a non-empty grid position is visible to the agent
+        """
+
+        coordinates = self.relative_coords(x, y)
+        if coordinates is None:
+            return False
+        vx, vy = coordinates
+
+        obs = self.gen_obs()
+        obs_grid, _ = Grid.decode(obs['image'])
+        obs_cell = obs_grid.get(vx, vy)
+        world_cell = self.grid.get(x, y)
+
+        return obs_cell is not None and obs_cell.type == world_cell.type
+
+    def gen_obs_grid(self):
+        """
+        Generate the sub-grid observed by the agent.
+        This method also outputs a visibility mask telling us which grid
+        cells the agent can actually see.
+        """
+
+        topX, topY, botX, botY = self.get_view_exts()
+
+        grid = self.grid.slice(topX, topY, self.agent_view_size, self.agent_view_size)
+
+        for i in range(self.agent_dir + 1):
+            grid = grid.rotate_left()
+
+        # Process occluders and visibility
+        # Note that this incurs some performance cost
+        if not self.see_through_walls:
+            vis_mask = grid.process_vis(agent_pos=(self.agent_view_size // 2, self.agent_view_size - 1))
+        else:
+            vis_mask = np.ones(shape=(grid.width, grid.height), dtype=np.bool)
+
+        # Make it so the agent sees what it's carrying
+        # We do this by placing the carried object at the agent's position
+        # in the agent's partially observable view
+        agent_pos = grid.width // 2, grid.height - 1
+        if self.carrying:
+            grid.set(*agent_pos, self.carrying)
+        else:
+            grid.set(*agent_pos, None)
+
+        return grid, vis_mask
+
+    def gen_obs(self):
+        """
+        Generate the agent's view (partially observable, low-resolution encoding)
+        """
+
+        grid, vis_mask = self.gen_obs_grid()
+
+        # Encode the partially observable view into a numpy array
+        image = grid.encode(vis_mask)
+
+        assert hasattr(self, 'mission'), "environments must define a textual mission string"
+
+        # Observations are dictionaries containing:
+        # - an image (partially observable view of the environment)
+        # - the agent's direction/orientation (acting as a compass)
+        # - a textual mission string (instructions for the agent)
+        obs = {
+            'image': image,
+            'direction': self.agent_dir,
+            'mission': self.mission
+        }
+
+        return obs
+
+    def get_obs_render(self, obs, tile_size=TILE_PIXELS // 2):
+        """
+        Render an agent observation for visualization
+        """
+
+        grid, vis_mask = Grid.decode(obs)
+
+        # Render the whole grid
+        img = grid.render(
+            tile_size,
+            self.breadcrumbs,
+            agent_pos=(self.agent_view_size // 2, self.agent_view_size - 1),
+            agent_dir=3,
+            highlight_mask=vis_mask
+        )
+
+        return img
+
+    @property
+    def unwrapped(self):
+        """Completely unwrap this env.
+
+        Returns:
+            gym.Env: The base non-wrapped gym.Env instance
+        """
+        return self
+
+    # def __str__(self):
+    #     if self.spec is None:
+    #         return '<{} instance>'.format(type(self).__name__)
+    #     else:
+    #         return '<{}<{}>>'.format(type(self).__name__, self.spec.id)
+
+    def __enter__(self):
+        """Support with-statement for the environment. """
+        return self
+
+    def __exit__(self, *args):
+        """Support with-statement for the environment. """
+        # self.close()
+        # propagate exception
+        return False
